feat(new-trace): Refactoring file structure.
