+++ static/app/views/performance/newTraceDetails/traceDrawer/details/span/sections/description.tsx
@@ -18,7 +18,7 @@ import {
 import {ModuleName} from 'sentry/views/starfish/types';
 import {resolveSpanModule} from 'sentry/views/starfish/utils/resolveSpanModule';
 
-import {TraceDrawerComponents} from '../../../../styles';
+import {TraceDrawerComponents} from '../../styles';
 
 export function SpanDescription({
   node,
