+++ static/app/views/performance/newTraceDetails/traceDrawer/details/span/sections/generalInfo.tsx
@@ -17,7 +17,7 @@ import {
   CardContentSubject,
   type SectionCardKeyValueList,
   TraceDrawerComponents,
-} from '../../../../styles';
+} from '../../styles';
 
 import {getSpanAncestryAndGroupingItems} from './ancestry';
